bug fixes
Fix vault integration not being use din several cases : /tan balance, salary payment and taxes
Fix vault successfully loaded lang file

+++ src/main/java/org/leralix/tan/commands/playersubcommand/SellRareItem.java
@@ -67,7 +67,7 @@ public class SellRareItem extends SubCommand {
         player.sendMessage(Lang.RARE_ITEM_SELLING_SUCCESS.get(quantity,
                 itemMeta.hasDisplayName() ? itemMeta.getDisplayName() : itemStack.getType().name(),
                 value * quantity));
-        EconomyUtil.addFromBalance(player, value * itemStack.getAmount());
+        EconomyUtil.addFromBalance(player, (double) value * itemStack.getAmount());
         player.getInventory().setItemInMainHand(new ItemStack(Material.AIR));
     }
 }
